dbms_project/requirements.txt
flask
flask_sqlalchemy
python-dotenv
google-generativeai
sqlalchemy
blinker
itsdangerous
jinja2
markupsafe
werkzeug
